+++ src/StateSmith/compiler/HistoryVertex.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 
 using StateSmith.compiler.Visitors;
+using System.Collections.Generic;
 
 namespace StateSmith.Compiling
 {
@@ -8,6 +9,16 @@ namespace StateSmith.Compiling
     {
         internal string stateTrackingVarName = "";
         internal int trackedStateCount = 0;
+
+        public override void Accept(VertexVisitor visitor)
+        {
+            visitor.Visit(this);
+        }
+    }
+
+    public class HistoryContinueVertex : PseudoStateVertex
+    {
+        internal List<ShallowHistoryVertex> historyVertices = new();
 
         public override void Accept(VertexVisitor visitor)
         {