+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## WIP
 ### Added
-- Add shallow history.  https://github.com/StateSmith/StateSmith/issues/63
+- Add history.  https://github.com/StateSmith/StateSmith/issues/63
+  - Deep history functionality supported via history continue nodes.
 - Add `$prefix` methods.
   - Experimental feature to help with duplicate state names. https://github.com/StateSmith/StateSmith/issues/65